turn image module into a subpackage, provide image.tiff moudle for tiff manipulation, bring rtc2color into image package
