+++ hyp3lib/image/rtc2color.py
